Fixed import issues and renamed the JobOfferDetailsService to JobOfferApplicationDetailsService and ProjectDetailsService to ProjectApplicationDetailsService
